Move mock profile into project-wide fixtures
